move deprecations to common
